refactor(plugin-swarm): rename callCollaborator action and tool
